agrego test y arreglo en el calculo de los tiempos

+++ TADP-tp2/src/ar/edu/tadp/viajes/ModuloTransporte.scala
@@ -25,7 +25,7 @@ object ModuloTransporte {
     return null
   }
 
-  def distanciaColectivo(linea: Int, dir1: Direccion, dir2: Direccion): Int = {
+  def distanciaColectivo(linea: Int, dir1: Direccion, dir2: Direccion): Double = {
     return 0
   }
